+++ src/arrays/objectives.ts
@@ -20,7 +20,11 @@ const NO_HIT_EXCEPTION_BOSSES = new ReadonlySet([
 export const NO_HIT_BOSSES: readonly BossID[] = BOSS_IDS.filter(
   (bossID) => !NO_HIT_EXCEPTION_BOSSES.has(bossID),
   // eslint-disable-next-line unicorn/prefer-spread
-).concat(BossIDCustom.KRAMPUS);
+).concat(
+  /// BossIDCustom.KRAMPUS, // (uncomment if another objective is needed)
+  BossIDCustom.URIEL,
+  BossIDCustom.GABRIEL,
+);
 
 export const ALL_OBJECTIVES: readonly Objective[] = (() => {
   const objectives: Objective[] = [];