+++ src/enums/OtherUnlockKind.ts
@@ -1,7 +1,9 @@
 export enum OtherUnlockKind {
   BEDS,
   SHOPKEEPERS,
-  BLUE_PURPLE_FIREPLACES,
+
+  BLUE_FIREPLACES,
+  PURPLE_FIREPLACES,
 
   GOLD_TRINKETS,
   GOLD_PILLS,
@@ -32,8 +34,12 @@ export function getOtherUnlockName(
       return ["entity", "Shopkeepers"];
     }
 
-    case OtherUnlockKind.BLUE_PURPLE_FIREPLACES: {
-      return ["entity", "Blue/Purple Fireplaces"];
+    case OtherUnlockKind.BLUE_FIREPLACES: {
+      return ["entity", "Blue Fireplaces"];
+    }
+
+    case OtherUnlockKind.PURPLE_FIREPLACES: {
+      return ["entity", "Purple Fireplaces"];
     }
 
     case OtherUnlockKind.GOLD_TRINKETS: {